backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
langchain_groq
langchain_community
yfinance
duckduckgo-search
ddgs
python-dotenv
requests
